+++ src/components/ChatBtn/ChatBtn.tsx
@@ -1,0 +1,17 @@
+import { useRecoilState } from 'recoil';
+import { modalIsOpenState } from 'recoil/atoms/commonAtoms';
+import * as styled from './ChatBtn.styles';
+
+export const ChatBtn = () => {
+  const [modalIsOpen, setModalIsOpen] = useRecoilState(modalIsOpenState);
+
+  const openModal = () => {
+    setModalIsOpen(true);
+  };
+
+  return (
+    <styled.ChatButton onClick={openModal} $isOpen={modalIsOpen}>
+      <styled.Icon />
+    </styled.ChatButton>
+  );
+};
