+++ src/components/ChatBtn/index.ts
@@ -1,0 +1,1 @@
+export { ChatBtn } from './ChatBtn';
