+++ src/components/Table/Table.constant.ts
@@ -1,15 +1,19 @@
+import ubit from 'assets/images/upbit.svg';
+import binance from 'assets/images/binance.svg';
+import bithumb from 'assets/images/bithumb.svg';
+
 export const SELECT_OPTION = [
   {
     key: 'upbit',
     value: 'upbit',
     title: '업비트',
-    symbol: 'https://kimpga.com/images/exchange-symbols/upbit.svg',
+    symbol: ubit,
   },
   {
     key: 'bithumb',
     value: 'bithumb',
     title: '빗썸',
-    symbol: 'https://kimpga.com/images/exchange-symbols/bithumb.svg',
+    symbol: bithumb,
   },
 ];
 
@@ -18,6 +22,6 @@ export const SELECT_SINGLE_OPTION = [
     key: 'binance',
     value: 'binance',
     title: '바이낸스 USDT 마켓',
-    symbol: 'https://kimpga.com/images/exchange-symbols/binance.svg',
+    symbol: binance,
   },
 ];
