+++ src/recoil/atoms/commonAtoms.ts
@@ -1,4 +1,4 @@
-import { atom } from 'recoil';
+import { atom, selector } from 'recoil';
 
 export const baseExchangeState = atom<string>({
   key: 'baseExchangeState',
@@ -8,6 +8,14 @@ export const baseExchangeState = atom<string>({
 export const modalIsOpenState = atom<boolean>({
   key: 'modalIsOpenState',
   default: false,
+});
+
+export const modalIsOpenSelector = selector({
+  key: 'modalIsOpenSelector',
+  get: ({ get }) => {
+    const modalIsOpen = get(modalIsOpenState);
+    return modalIsOpen;
+  },
 });
 
 export const marketCodesState = atom<string[]>({
